pyproject.toml
[project]
name = "brain-os-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.0",
    "fastapi>=0.117.1",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
